Move thunk to books
