+++ src/components/book-list.js
@@ -1,10 +1,10 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import Book from './one-book';
-import { fetchBooks } from '../redux/thunk/thunk';
+import { fetchBooks } from '../redux/books/books';
 
 const BooksList = () => {
-  const bookList = useSelector((state) => state.booksList);
+  const bookList = useSelector((state) => state.books);
   const dispatch = useDispatch();
 
   useEffect(() => {
