+++ src/redux/configureStore.js
@@ -2,11 +2,9 @@ import { createStore, combineReducers, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
 import booksReducer from './books/books';
-import thunkReducer from './thunk/thunk';
 
 const reducer = combineReducers({
   books: booksReducer,
-  booksList: thunkReducer,
 });
 
 const store = createStore(
